Move answer sound handler to an import
